# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "integr8scode",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/executions/{execution_id}/events": {
      "get": {
        "summary": "Get Execution Events",
        "description": "Get all events for an execution.",
        "operationId": "get_execution_events_api_v1_executions__execution_id__events_get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          },
          {
            "name": "event_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventType"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Event types to filter",
              "title": "Event Types"
            },
            "description": "Event types to filter"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExecutionEventResponse"
                  },
                  "title": "Response Get Execution Events Api V1 Executions  Execution Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events/executions/{execution_id}/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Execution Events",
        "operationId": "get_execution_events_api_v1_events_executions__execution_id__events_get",
        "parameters": [
          {
            "name": "execution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Execution Id"
            }
          },
          {
            "name": "include_system_events",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include system-generated events",
              "default": false,
              "title": "Include System Events"
            },
            "description": "Include system-generated events"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/user": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get User Events",
        "description": "Get events for the current user",
        "operationId": "get_user_events_api_v1_events_user_get",
        "parameters": [
          {
            "name": "event_types",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Types"
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/query": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Query Events",
        "operationId": "query_events_api_v1_events_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventFilterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/correlation/{correlation_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Events By Correlation",
        "operationId": "get_events_by_correlation_api_v1_events_correlation__correlation_id__get",
        "parameters": [
          {
            "name": "correlation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Correlation Id"
            }
          },
          {
            "name": "include_all_users",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include events from all users (admin only)",
              "default": false,
              "title": "Include All Users"
            },
            "description": "Include events from all users (admin only)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/current-request": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Current Request Events",
        "operationId": "get_current_request_events_api_v1_events_current_request_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/statistics": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event Statistics",
        "operationId": "get_event_statistics_api_v1_events_statistics_get",
        "parameters": [
          {
            "name": "start_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start time for statistics (defaults to 24 hours ago)",
              "title": "Start Time"
            },
            "description": "Start time for statistics (defaults to 24 hours ago)"
          },
          {
            "name": "end_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End time for statistics (defaults to now)",
              "title": "End Time"
            },
            "description": "End time for statistics (defaults to now)"
          },
          {
            "name": "include_all_users",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include stats from all users (admin only)",
              "default": false,
              "title": "Include All Users"
            },
            "description": "Include stats from all users (admin only)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get Event",
        "description": "Get a specific event by ID",
        "operationId": "get_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "events"
        ],
        "summary": "Delete Event",
        "operationId": "delete_event_api_v1_events__event_id__delete",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/publish": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Publish Custom Event",
        "operationId": "publish_custom_event_api_v1_events_publish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/events/aggregate": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Aggregate Events",
        "operationId": "aggregate_events_api_v1_events_aggregate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventAggregationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response Aggregate Events Api V1 Events Aggregate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventType": {
        "type": "string",
        "enum": [
          "execution_requested",
          "execution_accepted",
          "execution_queued",
          "execution_started",
          "execution_running",
          "execution_completed",
          "execution_failed",
          "execution_timeout",
          "execution_cancelled",
          "pod_created",
          "pod_scheduled",
          "pod_running",
          "pod_succeeded",
          "pod_failed",
          "pod_terminated",
          "pod_deleted",
          "user_registered",
          "user_login",
          "user_logged_in",
          "user_logged_out",
          "user_updated",
          "user_deleted",
          "user_settings_updated",
          "user_theme_changed",
          "user_notification_settings_updated",
          "user_editor_settings_updated",
          "notification_created",
          "notification_sent",
          "notification_delivered",
          "notification_failed",
          "notification_read",
          "notification_clicked",
          "notification_preferences_updated",
          "script_saved",
          "script_deleted",
          "script_shared",
          "security_violation",
          "rate_limit_exceeded",
          "auth_failed",
          "resource_limit_exceeded",
          "quota_exceeded",
          "system_error",
          "service_unhealthy",
          "service_recovered",
          "result_stored",
          "result_failed",
          "saga_started",
          "saga_completed",
          "saga_failed",
          "saga_cancelled",
          "saga_compensating",
          "saga_compensated",
          "create_pod_command",
          "delete_pod_command",
          "allocate_resources_command",
          "release_resources_command"
        ],
        "title": "EventType",
        "description": "Event types used throughout the system."
      },
...
      "ExecutionEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "$ref": "#/components/schemas/EventType"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "timestamp",
          "payload"
        ],
        "title": "ExecutionEventResponse",
        "description": "Model for execution event response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "description": "Sort order for queries."
      },
...
    }
  }
}